+++ auditlog_tests/tests.py
@@ -21,7 +21,7 @@ from auditlog.context import set_actor
 from auditlog.diff import model_instance_diff
 from auditlog.middleware import AuditlogMiddleware
 from auditlog.models import LogEntry
-from auditlog.registry import AuditlogModelRegistry, auditlog
+from auditlog.registry import AuditlogModelRegistry, AuditLogRegistrationError, auditlog
 from auditlog_tests.models import (
     AdditionalDataIncludedModel,
     AltPrimaryKeyModel,
@@ -35,6 +35,10 @@ from auditlog_tests.models import (
     PostgresArrayFieldModel,
     ProxyModel,
     RelatedModel,
+    SerializeNaturalKeyRelatedModel,
+    SerializeOnlySomeOfThisModel,
+    SerializePrimaryKeyRelatedModel,
+    SerializeThisModel,
     SimpleExcludeModel,
     SimpleIncludeModel,
     SimpleMappingModel,
@@ -1000,7 +1004,7 @@ class RegisterModelSettingsTest(TestCase):
 
         self.assertTrue(self.test_auditlog.contains(SimpleExcludeModel))
         self.assertTrue(self.test_auditlog.contains(ChoicesFieldModel))
-        self.assertEqual(len(self.test_auditlog.get_models()), 19)
+        self.assertEqual(len(self.test_auditlog.get_models()), 23)
 
     def test_register_models_register_model_with_attrs(self):
         self.test_auditlog._register_models(
@@ -1116,6 +1120,17 @@ class RegisterModelSettingsTest(TestCase):
         fields = self.test_auditlog.get_model_fields(SimpleExcludeModel)
         self.assertEqual(fields["include_fields"], ["label"])
         self.assertEqual(fields["exclude_fields"], ["text"])
+
+    def test_registration_error_if_bad_serialize_params(self):
+        with self.assertRaisesMessage(
+            AuditLogRegistrationError,
+            "Serializer options were given but the 'serialize_data' option is not "
+            "set. Did you forget to set serialized_data to True?",
+        ):
+            register = AuditlogModelRegistry()
+            register.register(
+                SimpleModel, serialize_kwargs={"fields": ["text", "integer"]}
+            )
 
 
 class ChoicesFieldModelTest(TestCase):
@@ -1533,4 +1548,226 @@ class ModelInstanceDiffTest(TestCase):
             changes,
             {"boolean": ("True", "False")},
             msg="ObjectDoesNotExist should be handled",
+        )
+
+
+class TestModelSerialization(TestCase):
+    def setUp(self):
+        super().setUp()
+        self.test_date = datetime.datetime(2022, 1, 1, 12, tzinfo=datetime.timezone.utc)
+        self.test_date_string = datetime.datetime.strftime(
+            self.test_date, "%Y-%m-%dT%XZ"
+        )
+
+    def test_does_not_serialize_data_when_not_configured(self):
+        instance = SimpleModel.objects.create(
+            text="sample text here", boolean=True, integer=4
+        )
+
+        log = instance.history.first()
+        self.assertIsNone(log.serialized_data)
+
+    def test_serializes_data_on_create(self):
+        with freezegun.freeze_time(self.test_date):
+            instance = SerializeThisModel.objects.create(
+                label="test label",
+                timestamp=self.test_date,
+                nullable=4,
+                nested={"foo": True, "bar": False},
+            )
+
+        log = instance.history.first()
+        self.assertTrue(isinstance(log, LogEntry))
+        self.assertEqual(log.action, 0)
+        self.assertDictEqual(
+            log.serialized_data["fields"],
+            {
+                "label": "test label",
+                "timestamp": self.test_date_string,
+                "nullable": 4,
+                "nested": {"foo": True, "bar": False},
+                "mask_me": None,
+                "date": None,
+                "code": None,
+            },
+        )
+
+    def test_serializes_data_on_update(self):
+        with freezegun.freeze_time(self.test_date):
+            instance = SerializeThisModel.objects.create(
+                label="test label",
+                timestamp=self.test_date,
+                nullable=4,
+                nested={"foo": True, "bar": False},
+            )
+
+        update_date = self.test_date + datetime.timedelta(days=4)
+        with freezegun.freeze_time(update_date):
+            instance.label = "test label change"
+            instance.save()
+
+        log = instance.history.filter(timestamp=update_date).first()
+        self.assertTrue(isinstance(log, LogEntry))
+        self.assertEqual(log.action, 1)
+        self.assertDictEqual(
+            log.serialized_data["fields"],
+            {
+                "label": "test label change",
+                "timestamp": self.test_date_string,
+                "nullable": 4,
+                "nested": {"foo": True, "bar": False},
+                "mask_me": None,
+                "date": None,
+                "code": None,
+            },
+        )
+
+    def test_serializes_data_on_delete(self):
+        with freezegun.freeze_time(self.test_date):
+            instance = SerializeThisModel.objects.create(
+                label="test label",
+                timestamp=self.test_date,
+                nullable=4,
+                nested={"foo": True, "bar": False},
+            )
+
+        obj_id = int(instance.id)
+        delete_date = self.test_date + datetime.timedelta(days=4)
+        with freezegun.freeze_time(delete_date):
+            instance.delete()
+
+        log = LogEntry.objects.filter(object_id=obj_id, timestamp=delete_date).first()
+        self.assertTrue(isinstance(log, LogEntry))
+        self.assertEqual(log.action, 2)
+        self.assertDictEqual(
+            log.serialized_data["fields"],
+            {
+                "label": "test label",
+                "timestamp": self.test_date_string,
+                "nullable": 4,
+                "nested": {"foo": True, "bar": False},
+                "mask_me": None,
+                "date": None,
+                "code": None,
+            },
+        )
+
+    def test_serialize_string_representations(self):
+        with freezegun.freeze_time(self.test_date):
+            instance = SerializeThisModel.objects.create(
+                label="test label",
+                nullable=4,
+                nested={"foo": 10, "bar": False},
+                timestamp="2022-03-01T12:00Z",
+                date="2022-04-05",
+                code="e82d5e53-ca80-4037-af55-b90752326460",
+            )
+
+        log = instance.history.first()
+        self.assertTrue(isinstance(log, LogEntry))
+        self.assertEqual(log.action, 0)
+        self.assertDictEqual(
+            log.serialized_data["fields"],
+            {
+                "label": "test label",
+                "timestamp": "2022-03-01T12:00:00Z",
+                "date": "2022-04-05",
+                "code": "e82d5e53-ca80-4037-af55-b90752326460",
+                "nullable": 4,
+                "nested": {"foo": 10, "bar": False},
+                "mask_me": None,
+            },
+        )
+
+    def test_serialize_mask_fields(self):
+        with freezegun.freeze_time(self.test_date):
+            instance = SerializeThisModel.objects.create(
+                label="test label",
+                nullable=4,
+                timestamp=self.test_date,
+                nested={"foo": 10, "bar": False},
+                mask_me="confidential",
+            )
+
+        log = instance.history.first()
+        self.assertTrue(isinstance(log, LogEntry))
+        self.assertEqual(log.action, 0)
+        self.assertDictEqual(
+            log.serialized_data["fields"],
+            {
+                "label": "test label",
+                "timestamp": self.test_date_string,
+                "nullable": 4,
+                "nested": {"foo": 10, "bar": False},
+                "mask_me": "******ential",
+                "date": None,
+                "code": None,
+            },
+        )
+
+    def test_serialize_only_auditlog_fields(self):
+        with freezegun.freeze_time(self.test_date):
+            instance = SerializeOnlySomeOfThisModel.objects.create(
+                this="this should be there", not_this="leave this bit out"
+            )
+
+        log = instance.history.first()
+        self.assertTrue(isinstance(log, LogEntry))
+        self.assertEqual(log.action, 0)
+        self.assertDictEqual(
+            log.serialized_data["fields"], {"this": "this should be there"}
+        )
+        self.assertDictEqual(
+            log.changes_dict,
+            {"this": ["None", "this should be there"], "id": ["None", "1"]},
+        )
+
+    def test_serialize_related(self):
+        with freezegun.freeze_time(self.test_date):
+            serialize_this = SerializeThisModel.objects.create(
+                label="test label",
+                nested={"foo": "bar"},
+                timestamp=self.test_date,
+            )
+            instance = SerializePrimaryKeyRelatedModel.objects.create(
+                serialize_this=serialize_this,
+                subheading="use a primary key for this serialization, please.",
+                value=10,
+            )
+
+        log = instance.history.first()
+        self.assertTrue(isinstance(log, LogEntry))
+        self.assertEqual(log.action, 0)
+        self.assertDictEqual(
+            log.serialized_data["fields"],
+            {
+                "serialize_this": serialize_this.id,
+                "subheading": "use a primary key for this serialization, please.",
+                "value": 10,
+            },
+        )
+
+    def test_serialize_related_with_kwargs(self):
+        with freezegun.freeze_time(self.test_date):
+            serialize_this = SerializeThisModel.objects.create(
+                label="test label",
+                nested={"foo": "bar"},
+                timestamp=self.test_date,
+            )
+            instance = SerializeNaturalKeyRelatedModel.objects.create(
+                serialize_this=serialize_this,
+                subheading="use a natural key for this serialization, please.",
+                value=11,
+            )
+
+        log = instance.history.first()
+        self.assertTrue(isinstance(log, LogEntry))
+        self.assertEqual(log.action, 0)
+        self.assertDictEqual(
+            log.serialized_data["fields"],
+            {
+                "serialize_this": "test label",
+                "subheading": "use a natural key for this serialization, please.",
+                "value": 11,
+            },
         )